Add files via upload
This is an adoption center for pets, one can browse the available pets as well as post their own pets for others to adopt. Whether someone is looking to give their pet a new home or hoping to welcome a new companion into their life, the platform offers a simple and user-friendly interface to connect pet owners and adopters. Each pet listing includes details such as name, age, breed, and description, helping ensure pets find the right homes. The goal is to make the adoption process easier, more accessible, and filled with compassion.